so clooose to printing higscores

+++ lib/CAN/CAN.h
@@ -20,7 +20,7 @@
 
 
 typedef struct {
-	unsigned int id;
+	uint16_t id;
 	uint8_t length;
 	uint8_t data[8];
 }can_message_t;
